feat & refactor : toast notification (error) 받는 배열 추가 & useToast생성 후 코드 분리
